Deal with corrupt rdb file issue
When `pkgbuild` is updated because the base docker image
has an outdated version and is then used to install a package,
the error pkgbuild.rdb is corrupt comes up. By updating packages
in a separate R session from where the git checkout package is 
built, we can (usually) avoid the problem.